now uses Principal in AccountController

+++ tenmo-server/src/main/java/com/techelevator/tenmo/dao/AccountDao.java
@@ -6,6 +6,6 @@ import java.math.BigDecimal;
 
 public interface AccountDao {
 
-    public Account getAccountByUserId(Long userId);
+    public Account getAccountByUserId(int userId);
 
 }
